python/requirements.txt
pyserial>=3.5
dataclasses; python_version<'3.7'